Move HaveHelpFlag out of flag_parser
It's only used by the cvd command, so moved it there.
