+++ base/cvd/cuttlefish/flag_parser/flag_parser.cpp
@@ -49,18 +49,6 @@ namespace cuttlefish {
 
 namespace {
 
-/*
- * From external/gflags/src, commit:
- *  061f68cd158fa658ec0b9b2b989ed55764870047
- *
- */
-constexpr std::array help_bool_opts{
-    "help", "helpfull", "helpshort", "helppackage", "helpxml", "version", "h"};
-constexpr std::array help_str_opts{
-    "helpon",
-    "helpmatch",
-};
-
 bool ShouldBeNullOpt(std::string_view value, CoerceToNullopt opt) {
   switch (opt) {
     case CoerceToNullopt::None:
@@ -778,23 +766,6 @@ Flag GflagsCompatFlag(const std::string& name,
                       std::optional<std::vector<unsigned>>& value,
                       CoerceToNullopt opt) {
   return GflagsCompatFlagImpl(name, value, opt);
-}
-
-Result<bool> HasHelpFlag(const std::vector<std::string>& args) {
-  std::vector<std::string> copied_args(args);
-  std::vector<Flag> flags;
-  flags.reserve(help_bool_opts.size() + help_str_opts.size());
-  bool bool_value_placeholder = false;
-  std::string str_value_placeholder;
-  for (const auto bool_opt : help_bool_opts) {
-    flags.emplace_back(GflagsCompatFlag(bool_opt, bool_value_placeholder));
-  }
-  for (const auto str_opt : help_str_opts) {
-    flags.emplace_back(GflagsCompatFlag(str_opt, str_value_placeholder));
-  }
-  CF_EXPECT(ConsumeFlags(flags, copied_args));
-  // if there was any match, some in copied_args were consumed.
-  return (args.size() != copied_args.size());
 }
 
 }  // namespace cuttlefish
