+++ base/cvd/cuttlefish/host/commands/cvd/cli/command_request.cpp
@@ -23,13 +23,27 @@
 
 #include <android-base/file.h>
 
-#include "cuttlefish/flag_parser/flag_parser.h"
 #include "cuttlefish/host/commands/cvd/cli/command_request.h"
 #include "cuttlefish/host/commands/cvd/cli/selector/selector_common_parser.h"
 #include "cuttlefish/result/result.h"
 
 namespace cuttlefish {
 
+namespace {
+
+/*
+ * From external/gflags/src, commit:
+ *  061f68cd158fa658ec0b9b2b989ed55764870047
+ *
+ */
+constexpr std::array help_bool_opts{
+    "help", "helpfull", "helpshort", "helppackage", "helpxml", "version", "h"};
+constexpr std::array help_str_opts{
+    "helpon",
+    "helpmatch",
+};
+
+}  // namespace
 CommandRequest::CommandRequest(cvd_common::Args args, cvd_common::Envs env,
                                selector::SelectorOptions selectors)
     : args_(std::move(args)),
@@ -91,6 +105,23 @@ CommandRequestBuilder CommandRequestBuilder::AddEnvVar(std::string key,
 Result<CommandRequest> CommandRequestBuilder::Build() && {
   return CommandRequest(std::move(args_), std::move(env_),
                         std::move(selector_options_));
+}
+
+Result<bool> HasHelpFlag(const std::vector<std::string>& args) {
+  std::vector<std::string> copied_args(args);
+  std::vector<Flag> flags;
+  flags.reserve(help_bool_opts.size() + help_str_opts.size());
+  bool bool_value_placeholder = false;
+  std::string str_value_placeholder;
+  for (const auto bool_opt : help_bool_opts) {
+    flags.emplace_back(GflagsCompatFlag(bool_opt, bool_value_placeholder));
+  }
+  for (const auto str_opt : help_str_opts) {
+    flags.emplace_back(GflagsCompatFlag(str_opt, str_value_placeholder));
+  }
+  CF_EXPECT(ConsumeFlags(flags, copied_args));
+  // if there was any match, some in copied_args were consumed.
+  return (args.size() != copied_args.size());
 }
 
 }  // namespace cuttlefish
