+++ base/cvd/cuttlefish/host/commands/cvd/cvd.cpp
@@ -21,7 +21,6 @@
 #include <unordered_map>
 #include <vector>
 
-#include "cuttlefish/flag_parser/flag_parser.h"
 #include "cuttlefish/host/commands/cvd/cli/command_request.h"
 #include "cuttlefish/host/commands/cvd/cli/frontline_parser.h"
 #include "cuttlefish/host/commands/cvd/cli/request_context.h"
